add func

+++ src/main/java/com/coolvetclinicpumb/vetclinicapp/dto/CategoryResponseDto.java
@@ -1,5 +1,5 @@
 package com.coolvetclinicpumb.vetclinicapp.dto;
 
-public record CategoryResponseDto(Long id,
+public record CategoryResponseDto(Short id,
                                   String typeCategory) {
 }
